feat(usm): Separate USM into two components
