--- src/features/unified-share-modal/UnifiedShareModal.stories.md
@@ -1,1 +1,0 @@
-`import UnifiedShareModal from 'box-ui-elements/es/features/unified-share-modal/UnifiedShareModal';`
