+++ src/features/unified-share-modal/stories/UnifiedShareModal.stories.js
@@ -1,11 +1,12 @@
 // @flow
+/* eslint-disable no-console */
 import * as React from 'react';
 import { IntlProvider } from 'react-intl';
 import { State, Store } from '@sambego/storybook-state';
 
-import Button from '../../components/button/Button';
+import Button from '../../../components/button/Button';
 
-import UnifiedShareModal from './UnifiedShareModal';
+import UnifiedShareModal from '../UnifiedShareModal';
 import notes from './UnifiedShareModal.stories.md';
 
 // Base Example. Extend for different initial loads or to demonstrate different interactions
@@ -22,6 +23,7 @@ const DEFAULT_SHARED_LINK_STATE = {
 };
 
 const INITIAL_STATE = {
+    isConfirmModalOpen: false,
     isOpen: false,
     item: {
         bannerPolicy: {
@@ -748,6 +750,172 @@ export const withAutofocusedSharedLink = () => {
                             Open USM Modal
                         </Button>
                     </div>
+                </IntlProvider>
+            )}
+        </State>
+    );
+};
+
+export const withFormOnly = () => {
+    const componentStore = createComponentStore();
+    const fakeRequest = () => {
+        // submitting is used to disable input fields, and not to show the loading indicator
+        componentStore.set({ submitting: true });
+        return new Promise(resolve => {
+            setTimeout(() => {
+                componentStore.set({ submitting: false });
+                resolve();
+            }, 500);
+        });
+    };
+
+    const getInitialData = () => {
+        const initialPromise = fakeRequest();
+        const fetchCollaborators = new Promise(resolved => {
+            setTimeout(() => {
+                const collaborators = contacts.slice();
+
+                const collaboratorsList = {
+                    collaborators,
+                };
+                componentStore.set({ collaboratorsList });
+                resolved();
+            }, 1000);
+        });
+        return Promise.all([initialPromise, fetchCollaborators]);
+    };
+
+    return (
+        <State store={componentStore}>
+            {state => (
+                <IntlProvider locale="en">
+                    <UnifiedShareModal
+                        canInvite
+                        changeSharedLinkAccessLevel={newLevel => {
+                            return fakeRequest().then(() => {
+                                return componentStore.set({
+                                    sharedLink: {
+                                        ...state.sharedLink,
+                                        accessLevel: newLevel,
+                                    },
+                                });
+                            });
+                        }}
+                        changeSharedLinkPermissionLevel={newLevel => {
+                            return fakeRequest().then(() => {
+                                return componentStore.set({
+                                    sharedLink: {
+                                        ...state.sharedLink,
+                                        permissionLevel: newLevel,
+                                    },
+                                });
+                            });
+                        }}
+                        collaboratorsList={state.collaboratorsList}
+                        collaborationRestrictionWarning="Collaboration invitations can only be sent to people within Box Corporate"
+                        currentUserID="0"
+                        getCollaboratorContacts={() => {
+                            return Promise.resolve(contacts);
+                        }}
+                        getSharedLinkContacts={() => {
+                            return Promise.resolve(contacts);
+                        }}
+                        getInitialData={getInitialData}
+                        inviteePermissions={[
+                            { default: false, text: 'Co-owner', value: 'Co-owner' },
+                            { default: true, text: 'Editor', value: 'Editor' },
+                            { default: false, text: 'Viewer Uploader', value: 'Viewer Uploader' },
+                            { default: false, text: 'Previewer Uploader', value: 'Previewer Uploader' },
+                            { default: false, text: 'Viewer', value: 'Viewer' },
+                            { default: false, text: 'Previewer', value: 'Previewer' },
+                            { default: false, text: 'Uploader', value: 'Uploader' },
+                        ]}
+                        isOpen={state.isOpen}
+                        isToggleEnabled
+                        item={state.item}
+                        onAddLink={() => {
+                            fakeRequest().then(() => {
+                                componentStore.set({
+                                    sharedLink: {
+                                        accessLevel: 'peopleInYourCompany',
+                                        allowedAccessLevels: {
+                                            peopleWithTheLink: true,
+                                            peopleInYourCompany: true,
+                                            peopleInThisItem: true,
+                                        },
+                                        canChangeAccessLevel: true,
+                                        enterpriseName: 'Box',
+                                        expirationTimestamp: 1509173940,
+                                        isDownloadSettingAvailable: true,
+                                        isNewSharedLink: true,
+                                        permissionLevel: 'canViewDownload',
+                                        url: 'https://box.com/s/abcdefg',
+                                    },
+                                });
+                            });
+                        }}
+                        onRemoveLink={() => {
+                            fakeRequest().then(() => {
+                                componentStore.set({
+                                    sharedLink: DEFAULT_SHARED_LINK_STATE,
+                                });
+                                console.log('removed link');
+                            });
+                        }}
+                        /* eslint-disable-next-line no-alert */
+                        onSettingsClick={() => alert('hi!')}
+                        recommendedSharingTooltipCalloutName=""
+                        sendInvites={() =>
+                            fakeRequest().then(() => {
+                                console.log('sent invites');
+                            })
+                        }
+                        sendInvitesError=""
+                        sendSharedLink={() =>
+                            fakeRequest().then(() => {
+                                console.log('sent shared link');
+                            })
+                        }
+                        sendSharedLinkError=""
+                        sharedLink={state.sharedLink}
+                        showCalloutForUser
+                        showFormOnly
+                        showUpgradeOptions
+                        submitting={state.submitting}
+                        suggestedCollaborators={{
+                            '2': {
+                                id: '2',
+                                userScore: 0.1,
+                                name: 'David',
+                                email: 'dt@example.com',
+                                type: 'user',
+                            },
+                            '5': {
+                                id: '5',
+                                userScore: 0.2,
+                                name: 'Will',
+                                email: 'wy@example.com',
+                                type: 'user',
+                            },
+                            '1': {
+                                id: '1',
+                                userScore: 0.5,
+                                name: 'Jeff',
+                                email: 'jt@example.com',
+                                type: 'user',
+                            },
+                            '3': { id: '3', userScore: 2, name: 'Yang', email: 'yz@example.com', type: 'user' },
+                        }}
+                        trackingProps={{
+                            collaboratorListTracking: {},
+                            inviteCollabsEmailTracking: {},
+                            inviteCollabTracking: {},
+                            modalTracking: {},
+                            removeLinkConfirmModalTracking: {},
+                            sharedLinkEmailTracking: {},
+                            sharedLinkTracking: {},
+                        }}
+                    />
                 </IntlProvider>
             )}
         </State>
