+++ src/features/unified-share-modal/stories/UnifiedShareModal.stories.md
@@ -1,0 +1,3 @@
+`import UnifiedShareModal from 'box-ui-elements/es/features/unified-share-modal/UnifiedShareModal';`
+
+The `UnifiedShareModal` includes a modal wrapper with a form (the `UnifiedShareForm`) by default. To render the form without the wrapper, set the `showFormOnly` prop to `true`.
